# docs/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "Claw-Empire API",
    "version": "1.2.0",
    "description": "Baseline OpenAPI contract for commonly used Claw-Empire endpoints."
  },
  "servers": [
    {
      "url": "/",
      "description": "Current origin (recommended for local docs)"
    },
    {
      "url": "http://127.0.0.1:8790",
      "description": "Local development (loopback)"
    },
    {
      "url": "http://localhost:8790",
      "description": "Local development (localhost)"
    }
  ],
...
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "API_AUTH_TOKEN"
      },
      "inboxSecret": {
        "type": "apiKey",
        "in": "header",
        "name": "x-inbox-secret"
      }
    },
    "schemas": {
      "ErrorResponse": {
        "type": "object",
        "properties": {
          "error": {
            "type": "string"
          },
          "message": {
            "type": "string"
          }
        },
        "required": ["error"],
        "additionalProperties": true
      },
...
    }
  },
  "paths": {
...
    "/api/auth/session": {
      "get": {
        "tags": ["runtime"],
        "summary": "Bootstrap authenticated session cookie",
        "responses": {
          "200": {
            "description": "Session ready"
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/inbox": {
      "post": {
        "tags": ["messages"],
        "summary": "External webhook inbox",
        "security": [
          {
            "inboxSecret": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "additionalProperties": true
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Webhook accepted"
          },
          "401": {
            "description": "Missing or invalid x-inbox-secret",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
...
  }
}